[Data][Fix] Adding a column to pandas dataset (#51794)
<!-- Thank you for your contribution! Please review
https://github.com/ray-project/ray/blob/master/CONTRIBUTING.rst before
opening a pull request. -->

<!-- Please add a reviewer to the assignee section when you create a PR.
If you don't have the access to it, we will shortly find a reviewer and
assign them to your PR. -->

## Why are these changes needed?
The index of the column must be set to align with the index of the batch
when adding a new column to a pandas dataset. Otherwise it will result
in a `Nan`.
<!-- Please give a short summary of the change and the problem this
solves. -->

## Related issue number
Closes #51758
<!-- For example: "Closes #1234" -->

## Checks

- [x] I've signed off every commit(by using the -s flag, i.e., `git
commit -s`) in this PR.
- [x] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for
https://docs.ray.io/en/master/.
- [ ] I've added any new APIs to the API Reference. For example, if I
added a
method in Tune, I've added it in `doc/source/tune/api/` under the
           corresponding `.rst` file.
- [x] I've made sure the tests are passing. Note that there might be a
few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [x] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(

---------

Signed-off-by: zhaoch23 <c233zhao@uwaterloo.ca>
Signed-off-by: zhilong <zhilong.chen@mail.mcgill.ca>
Co-authored-by: zhilong <zhilong.chen@mail.mcgill.ca>