[ray.data.llm] Add retries and exponential backoff to `HTTPRequestStage` (#52084)
<!-- Thank you for your contribution! Please review
https://github.com/ray-project/ray/blob/master/CONTRIBUTING.rst before
opening a pull request. -->

<!-- Please add a reviewer to the assignee section when you create a PR.
If you don't have the access to it, we will shortly find a reviewer and
assign them to your PR. -->

## Why are these changes needed?

Currently, `HTTPRequestStage` provides a very basic implementation of
sending a batch of requests and yielding responses one by one. However,
this can fail if the server is overwhelmed or for time out with very
long running requests.

This PR adds support for exponential backoff and configurable session
timeouts.

Currently, if we realize that a request fails, I retry them
individually. A potential improvement is to somehow batch failed
requests together and retry. However, that is non-trivial and I'm not
sure how beneficial that is. I'm leaving that as a future optimization.

<!-- Please give a short summary of the change and the problem this
solves. -->

## Related issue number

<!-- For example: "Closes #1234" -->

## Checks

- [ ] I've signed off every commit(by using the -s flag, i.e., `git
commit -s`) in this PR.
- [ ] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for
https://docs.ray.io/en/master/.
- [ ] I've added any new APIs to the API Reference. For example, if I
added a
method in Tune, I've added it in `doc/source/tune/api/` under the
           corresponding `.rst` file.
- [x] I've made sure the tests are passing. Note that there might be a
few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [x] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(

---------

Signed-off-by: sumanthrh <sumanthrh@anyscale.com>