Fix compatibility with GCP IAM conditions
If the Ray SA has been granted the required roles, but with conditions
(ie to restrict it to a particular subnet or storage buckets), the v1
policy will include a `withcond` in the role names, which breaks the
comparison. From version 3 the policy name is preserved and the
conditions are included in the response.

Ref https://cloud.google.com/iam/docs/troubleshooting-withcond#solution

Signed-off-by: Tarjei Husøy <git@thusoy.com>