[Data] Add `InMemorySizeEstimator` (#1607)
Towards https://github.com/anyscale/rayturbo/pull/1530

To implement `Dataset.load_from_uris`, we need a way to estimate the
in-memory size of additional input data. Since the logic for this
depends on the reader implementation, this PR introduces an abstraction
for estimating in-memory sizes and moves the logic out of
`plan_list_files`. A secondary motivation for this change is to avoid
adding too much complexity to the list files code.

---------

Signed-off-by: Balaji Veeramani <bveeramani@berkeley.edu>