[Data] Adding in separate instance var to track overhead for stream split iterator (#52436)
## Why are these changes needed?
In the existing code, we create a copy of the stats which 1. means that
the modified stats with the coordinator overhead is just a locally
modified copy and 2. this may cause issues with multithreading.

This code modifies the `SplitCoordinator` to store a overhead instance
variable and then when the stats are returned, we add the stored
instance variable to the returned stats which should solve both of the
above problems.


Signed-off-by: Matthew Owen <mowen@anyscale.com>