[serve][P0] fix await deployment response hanging issue (#52561)
## Why are these changes needed?

If the user passes a `DeploymentResponse` object directly into a
deployment handle, we will resolve it to the object ref under the hood
before scheduling the request. But we were incorrectly using a
threading.Lock when getting the object ref so this can cause deadlock if
the same `DeploymentResponse` object is used for multiple downstream
calls. Instead we should be using an asyncio lock.

## Related issue number

closes https://github.com/ray-project/ray/issues/51190

---------

Signed-off-by: Cindy Zhang <cindyzyx9@gmail.com>