[core] Deflake `test_runtime_env_3` (#52553)
The test is bumping up against the timeout which is causing flakes on
macos builds.

Rather than increase the timeout or split the test, I opted to refactor
it a bit to reduce the runtime by combining the logic in the tests to
avoid "source" and "option" parametrization. Also reduced to a single
worker node for each test to reduce overhead in CI.

I've also deleted some tests that have been dead code for a long time
(those testing the logic to disable GC, which hasn't existed for years).

---------

Signed-off-by: Edward Oakes <ed.nmi.oakes@gmail.com>