[train] Deserialize the user-defined training function directly on workers (#52684)
Serialize the user-defined training function to the object store on the
driver process, and deserialize the function along with any captured
imports directly on the train worker processes.

Previously, the function was passed directly to the controller process,
then again to the workers. This causes the train function to be
deserialized on the controller process, which adds a bunch of captured
modules which are unnecessary since the controller is not running user
code.

This PR also introduces a lightweight `ObjectRefWrapper` that is used to
manually dereference an object reference, since Ray Core's default
behavior is to dereference an object ref if passed to a Ray actor
constructor/method.

---------

Signed-off-by: Justin Yu <justinvyu@anyscale.com>
Signed-off-by: jhsu <jhsu@anyscale.com>