[data] Adding in batch inference mock pipeline release test (#52616)
## Why are these changes needed?
We want to add a release test that will be used to model performance on
a batch inference image pipeline. This is meant to more accurately model
a realistic user pipeline. This will be used to track our improvement on
this workload and more broadly our improvement on batch inference with
images with and without spot nodes.

## Related issue number

<!-- For example: "Closes #1234" -->

## Checks

- [ ] I've signed off every commit(by using the -s flag, i.e., `git
commit -s`) in this PR.
- [ ] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for
https://docs.ray.io/en/master/.
- [ ] I've added any new APIs to the API Reference. For example, if I
added a
method in Tune, I've added it in `doc/source/tune/api/` under the
           corresponding `.rst` file.
- [ ] I've made sure the tests are passing. Note that there might be a
few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [ ] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(

---------

Signed-off-by: Matthew Owen <mowen@anyscale.com>
Co-authored-by: Lonnie Liu <lonnie@anyscale.com>