[train] add XGBoostTrainer user guide (#52355)
This PR adds a comprehensive guide for distributed XGBoost training with
Ray Train. The guide follows the same structure as our other
framework-specific guides.

The PR also adds a deprecation notice to the older XGBoost/LightGBM
notebook and includes the guide in the main Train documentation
navigation.

In a follow-up PR, we will create a similar guide for LightGBM.

---------

Signed-off-by: Matthew Deng <matt@anyscale.com>
Signed-off-by: matthewdeng <matthew.j.deng@gmail.com>
Co-authored-by: Justin Yu <justinvyu@anyscale.com>