[core] Correctly release dynamic generator refs (#52843)
In https://github.com/ray-project/ray/pull/52095, the while loop in
DynamicObjectRefGenerator that pops the ref was just changed to a for
loop that iterates through the refs. This results in the refs sticking
around and using up memory when they can be destroyed.

Also changing the list to a deque for efficiency on popping from the
front.
---------

Signed-off-by: dayshah <dhyey2019@gmail.com>