[core] [docs] Dynamic generator deprecation (#52887)
Deprecating the dynamic ref generator.

It was supposed to be deprecated a long time ago in favor of streaming
generators but found that the deprecation warning on the docs page was
actually never showing
https://docs.ray.io/en/releases-2.46.0/ray-core/tasks/generators.html
because the warning is above the title of the page.

Moved the dynamic ref generator page under deprecated at the bottom of
the ray generators page and outside the tasks subsection.


Signed-off-by: dayshah <dhyey2019@gmail.com>