[ci] Remove `"serverless"` team CI steps (#52998)
There are no `team:serverless` tests in the repo at this point, so these
build steps appear to be doing nothing but burning instance time.

Also separated the spark-on-ray tests in core out into their own build
step (which is how it used to be for "serverless"). Planning to make
this postmerge-only.

---------

Signed-off-by: Edward Oakes <ed.nmi.oakes@gmail.com>