[ci] Make Ray client tests conditional (#53045)
- Makes Ray Client tests conditional on the `ray/util/client/` directory
- Pulls out the parametrized logic inside of the `test_placement_group*`
and `test_scheduling*` tests and instead uses the `RAY_CLIENT_MODE`
environment variable. There are no changes to these tests, the diff is
so large because they were all un-indented after removing the context
manager.

This should speed up premerge significantly (the placement group tests
are slow and this was ~doubling their runtime).

---------

Signed-off-by: Edward Oakes <ed.nmi.oakes@gmail.com>