[Dashboard] Fix dashboard for daylight savings (#52755)
Did not get traction with the [holistic
fix](https://github.com/ray-project/ray/pull/51314) for daylight savings
as there was pushback on usage of tzlocal in python. So, splitting out
the fix on the UI side which is perhaps less controversial.

On the UI side, `timezone.ts` is showing GMT offsets not reflective of
daylight savings time. Updating to dynamically pull the offset from
`Intl.DateTimeFormat`. Note not doing any dynamic evaluation so for
browser windows open during the switch, it will not update.

---------

Signed-off-by: Brendan Miller <6900229+bhmiller@users.noreply.github.com>
Co-authored-by: Alan Guo <aguo@anyscale.com>