[dashboard] Add a `/api/v0/logs/file` API param to customize the downloaded filename (#53008)
Add a configuration option to the log file streaming API to download it
to a specific filename on the client.

Before this PR, the downloaded file is just named `file`. The new
default is `file.txt` if nothing is provided.

Note: The `download_filename` is different from the existing `filename`
parameter. `filename` is used to *match* which file to download on the
server's log directory. `download_filename` is what tell the browser to
download the file as on the client side.

---------

Signed-off-by: Justin Yu <justinvyu@anyscale.com>