[Serve] Avoid concurrent modification of `LongPollClient.snapshot_ids` (#52793)
## Why are these changes needed?

This is a followup to our previous PR
https://github.com/ray-project/ray/pull/48807 , which changed the way
that the `LongPollClient` is managed by `Router`s such that there is, in
the long-term, only a single active `LongPollClient` per client process.

To make that happen, that PR added a way to change the set of
`snapshot_ids` that the `LongPollClient` was listening for.
Unfortunately, I didn't realize at the time that the registration code
that changes the `snapshot_ids` would be running in a different thread
from the `LongPollClient`'s polling loop, which means that the snapshot
ID dictionary could be concurrently mutated from one thread while the
other thread is trying to serialize it, leading to a stack trace like:

```
Exception in callback <function LongPollClient._process_update.<locals>.chained at 0x7f1d4afbc540>
handle: <Handle LongPollClient._process_update.<locals>.chained>
Traceback (most recent call last):
  File "uvloop/cbhandles.pyx", line 61, in uvloop.loop.Handle._run
  File ".../site-packages/ray/serve/_private/long_poll.py", line 199, in chained
    self._on_callback_completed(trigger_at=len(updates))
  File ".../site-packages/ray/serve/_private/long_poll.py", line 133, in _on_callback_completed
    self._poll_next()
  File ".../site-packages/ray/serve/_private/long_poll.py", line 143, in _poll_next
    self._current_ref = self.host_actor.listen_for_change.remote(self.snapshot_ids)
                        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File ".../site-packages/ray/actor.py", line 206, in remote
    return self._remote(args, kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^
  File ".../site-packages/ray/_private/auto_init_hook.py", line 21, in auto_init_wrapper
    return fn(*args, **kwargs)
           ^^^^^^^^^^^^^^^^^^^
  File ".../site-packages/ray/util/tracing/tracing_helper.py", line 422, in _start_span
    return method(self, args, kwargs, *_args, **_kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File ".../site-packages/ray/actor.py", line 366, in _remote
    return invocation(args, kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^
  File ".../site-packages/ray/actor.py", line 347, in invocation
    return actor._actor_method_call(
           ^^^^^^^^^^^^^^^^^^^^^^^^^
  File ".../site-packages/ray/actor.py", line 1503, in _actor_method_call
    object_refs = worker.core_worker.submit_actor_task(
                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "python/ray/_raylet.pyx", line 3972, in ray._raylet.CoreWorker.submit_actor_task
  File "python/ray/_raylet.pyx", line 3977, in ray._raylet.CoreWorker.submit_actor_task
  File "python/ray/_raylet.pyx", line 859, in ray._raylet.prepare_args_and_increment_put_refs
  File "python/ray/_raylet.pyx", line 850, in ray._raylet.prepare_args_and_increment_put_refs
  File "python/ray/_raylet.pyx", line 900, in ray._raylet.prepare_args_internal
  File "/app/virtualenv/lib/python3.12/site-packages/ray/_private/serialization.py", line 556, in serialize
    return self._serialize_to_msgpack(value)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File ".../site-packages/ray/_private/serialization.py", line 530, in _serialize_to_msgpack
    msgpack_data = MessagePackSerializer.dumps(value, _python_serializer)
                   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "python/ray/includes/serialization.pxi", line 175, in ray._raylet.MessagePackSerializer.dumps
  File ".../site-packages/msgpack/__init__.py", line 36, in packb
    return Packer(**kwargs).pack(o)
           ^^^^^^^^^^^^^^^^^^^^^^^^
  File "msgpack/_packer.pyx", line 279, in msgpack._cmsgpack.Packer.pack
  File "msgpack/_packer.pyx", line 276, in msgpack._cmsgpack.Packer.pack
  File "msgpack/_packer.pyx", line 265, in msgpack._cmsgpack.Packer._pack
  File "msgpack/_packer.pyx", line 211, in msgpack._cmsgpack.Packer._pack_inner
RuntimeError: dictionary changed size during iteration
```

If this happens (it's rare; we've been running the shared long poll code
for months but only noticed this when we started making a lot more Serve
deployments per unit time recently), the `LongPollClient` never calls
`poll_next()` again, and thus no longer receives updates from the Serve
Controller. Depending on how fast your actor replicas come and go, this
eventually leads to poor scheduling, and then no scheduling if all the
replicas you had at the time of the error go away.

A simple fix for this is to shallow copy the snapshot IDs just before
passing them to the `.remote()` call. The shallow copy is then not
modified if the original snapshot IDs held by the `LongPollClient` are
modified. We've tested this fix in our system and it did prevent the
problem from occurring.

Since it would be difficult to arrange this to happen in a deterministic
test, I've added a comment describing the problem case that the copy
solves for.

## Related issue number

<!-- For example: "Closes #1234" -->

## Checks

- [x] I've signed off every commit(by using the -s flag, i.e., `git
commit -s`) in this PR.
- [x] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for
https://docs.ray.io/en/master/.
- [ ] I've added any new APIs to the API Reference. For example, if I
added a
method in Tune, I've added it in `doc/source/tune/api/` under the
           corresponding `.rst` file.
- [ ] I've made sure the tests are passing. Note that there might be a
few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [ ] Unit tests
   - [ ] Release tests
   - [x] This PR is not tested :(

---------

Signed-off-by: Josh Karpel <josh.karpel@gmail.com>
Co-authored-by: Cindy Zhang <cindyzyx9@gmail.com>